requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
python-dotenv==1.0.1
schedule==1.2.1
lxml==5.1.0
flask==3.0.0
flask-cors==4.0.0
